chore: review cleanup — dead code removal, test fixes, arena hardening (#741)
* chore: review cleanup — remove dead code, fix tests, harden arena

- Remove compiled packc binary from git, add to .gitignore
- Fix schema validation test fragility (use local schemas)
- Remove dead SDK exports: CredentialOption, WithTurnDetector,
  WithStreamingVideo, empty hooks package
- Remove dead runtime types: tools.ToolPolicy, PredictionRequest,
  PredictionResponse, PredictMessage
- Remove deprecated arena statestore wrappers, update all callers
- Add context cancellation to chaos hook slow mode

* fix: revert SDK API removals — wire up, don't delete

Restore CredentialOption, WithTurnDetector, WithStreamingVideo, and
sdk/hooks package. These are public API surface that must not be
removed. They need wiring into pipeline/session/provider code paths,
tracked as backlog items.

* feat: wire up disconnected SDK options — credentials, turn detection, video

Wire three SDK options that were accepted but never consumed:

- WithCredential(...CredentialOption): resolves API keys from direct
  value, environment variable, or file with priority ordering. Feeds
  into resolveProvider() before provider detection.

- WithTurnDetector: passes custom turn detector through to the VAD
  pipeline's AudioTurnConfig for duplex streaming sessions.

- WithStreamingVideo: converts VideoStreamConfig to FrameRateLimitConfig
  and creates a FrameRateLimitStage in the pipeline when TargetFPS > 0.

All three include comprehensive unit tests.

* test: add integration tests for credential and video stream options

Add end-to-end integration tests using mock providers:

Credential tests (5): env var resolution, file-based credential,
direct API key, env var not set error, file not found error.

Video stream tests (2): streaming with frame rate limiter config,
unary Send with video config (verifies no pipeline breakage).

WithTurnDetector is covered by unit tests only — VAD mode requires
real STT/TTS services that can't be mocked in integration tests.

* feat: expose cloud credential overrides as PlatformOptions

Add PlatformOption variants that surface existing runtime credential
constructors through the SDK:

- WithAWSRoleARN: assume an IAM role for Bedrock access
- WithAWSProfile: use a named AWS credentials profile
- WithGCPServiceAccount: use a service account key file for Vertex
- WithAzureManagedIdentity: specify managed identity client ID
- WithAzureClientSecret: use service principal credentials

Also adds NewAWSCredentialWithProfile to the runtime credentials
package. Cloud credential resolvers moved to
platform_credential_integration.go (requires real cloud SDKs).

+++ arena/chaos/hook_test.go
@@ -103,7 +103,7 @@ func TestHook_BeforeExecution_SlowMode(t *testing.T) {
 	if !d.Allow {
 		t.Error("expected allow for slow mode (adds latency but doesn't block)")
 	}
-	if d.Metadata["chaos_delay_ms"] != 2000 {
+	if d.Metadata["chaos_delay_ms"] != int64(2000) {
 		t.Errorf("expected delay metadata, got %v", d.Metadata["chaos_delay_ms"])
 	}
 }
